Subtracao de quantidade no DB
Criação de classe que subtrai a quantidade no Banco de Dados e implementada no código do PDV
Concentração de strings repetitivas na classe BancoDeDadosControl.
Corrigida falha de que o preço não aceitava números decimais, mas só é aceito com ponto e não vírgula.

+++ PDV/formulario/src/main/java/com/propostapdv/formulario/ControllerFormulario.java
@@ -19,7 +19,7 @@ public class ControllerFormulario {
         }
 
         @PostMapping("/cadproduto")
-        public ModelAndView cadProduto(String nome, int preco, int quantidade) throws SQLException{
+        public ModelAndView cadProduto(String nome, double preco, int quantidade) throws SQLException{
             PDV.AddProduto(nome, preco, quantidade);
             ModelAndView mv = new ModelAndView("produtos");
             mv.addObject("listaDeProdutos", PDV.ListaProdutosStr);
